Style fixes, image handling change.
Minor style and bug fixes related to the way images display in both the
control page and search page.

Changed the way that images are referenced in both control and search
page. Added a more robust way to tell if an needs to be populated to the
image box, and added a generic image for those that don't.