Restructured image handling
Restructured some of the image handling to make it easier to get reuse
out of some of the functions. Also added a "loading" image placeholder
in the form of the "generic movie" icon. When I can find a better way to
handle .gif images in WPF I will use that instead.